[ARC-584] Reorganize package structure
branch: feature/ARC-584-clean-up-mpyl-example-and-try-to-use-it-as-the-main-runner
