+++ src/mpyl/reporting/formatting/__init__.py
@@ -1,0 +1,3 @@
+"""
+Pretty-print build results in various formats like `.md` and `.txt`
+"""
