+++ src/mpyl/reporting/formatting/text.py
@@ -2,10 +2,10 @@
 Simple run result formatters
 """
 
-from ..project import Stage
-from ..steps.models import Artifact, ArtifactType
-from ..steps.run import RunResult
-from ..utilities.junit import to_test_suites, sum_suites
+from ...project import Stage
+from ...steps.models import Artifact, ArtifactType
+from ...steps.run import RunResult
+from ...utilities.junit import to_test_suites, sum_suites
 
 
 def to_string(run_result: RunResult) -> str:
